Time: 0 ms (100.00%) | Memory: 23.5 MB (41.00%) - LeetSync

+++ 155-min-stack/min-stack.cpp
@@ -2,8 +2,8 @@
 
 class MinStack {
 public:
-    stack <long long> s;
-    stack <long long> min;
+    stack <int > s;
+    stack <int > min;
     MinStack() {
         min.push(INT_MAX);
     }
